% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kolada.R
\name{get_kpi_by_Ids}
\alias{get_kpi_by_Ids}
\title{Retrieve KPIs based on the filtered Kpi Ids}
\usage{
get_kpi_by_Ids(kpi_id)
}
\arguments{
\item{kpi_id}{Character vector. KPI IDs to retrieve.}
}
\value{
A tibble with columns:
\describe{
\item{id}{KPI ID}
\item{title}{KPI title}
\item{description}{KPI description}
\item{municipality_type}{Municipality type associated with the KPI}
\item{operating_area}{Operating area associated with the KPI}
}
}
\description{
Fetches KPIs from the Kolada API, combines paginated results, and returns a tibble.
}
\examples{
get_kpi_by_Ids(c("N00951", "U00002"))

}
