% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kolada.R
\name{get_municipality}
\alias{get_municipality}
\title{Retrieve municipality list from Kolada API}
\usage{
get_municipality()
}
\value{
A tibble with columns:
\describe{
\item{id}{Municipality ID}
\item{title}{Municipality title}
\item{type}{Municipality Type}
}
}
\description{
Fetches all municipality from the Kolada API, combines paginated results, and returns a tibble.
}
\examples{
municipality <- get_municipality()
head(municipality)

}
