% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kolada.R
\name{get_municipality_by_Ids}
\alias{get_municipality_by_Ids}
\title{Retrieve municipalities by its exact IDs}
\usage{
get_municipality_by_Ids(id)
}
\arguments{
\item{id}{Character vector. Municipality IDs to retrieve.}
}
\value{
A tibble with columns:
\describe{
\item{id}{Municipality ID}
\item{title}{Municipality title}
\item{type}{Municipality type}
}
}
\description{
Fetches municipalities from the Kolada API, combines paginated results, and returns a tibble.
}
\examples{
get_municipality_by_Ids(c("0001", "0003"))

}
