% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kolada.R
\name{kolada_base}
\alias{kolada_base}
\title{Get Kolada API base URL}
\usage{
kolada_base()
}
\value{
A character string containing the base URL of Kolada API
}
\description{
Returns the base URL of Kolada API
}
\examples{
kolada_base()

}
