% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kolada.R
\name{kolada_get_list}
\alias{kolada_get_list}
\title{GET request to Kolada API}
\usage{
kolada_get_list(path, query = list(), base = kolada_base())
}
\arguments{
\item{path}{Character string. API endpoint (e.g., "/kpi")}

\item{query}{List. Optional parameters to include in request}

\item{base}{Character string. Base URL of Kolada API}
}
\value{
A list containing the parsed JSON response
}
\description{
Sends a get request to a specified api endpoint and returns the parsed json to list.
}
\examples{
kolada_get_list("/kpi")

}
